+++ eotsmanager/cmd/eotsd/daemon/flags.go
@@ -17,7 +17,8 @@ const (
 	flagHDPath            = "hd-path"
 	flagIndex             = "index"
 	flagRecover           = "recover"
-	flagMnemonicSrc = "source"
+	flagMnemonicSrc       = "source"
+	passphraseFlag        = "passphrase"
 	flagDBPath      = "db-path"
 	flagBackupDir   = "backup-dir"
 )