+++ CHANGELOG.md
@@ -45,6 +45,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#449](https://github.com/babylonlabs-io/finality-provider/pull/449) chore: remove bbn address from eots keys - show and list
 * [#450](https://github.com/babylonlabs-io/finality-provider/pull/450) chore: metrics for votes
 * [#464](https://github.com/babylonlabs-io/finality-provider/pull/464) chore: fix flaky tests
+* [#469](https://github.com/babylonlabs-io/finality-provider/pull/469) feat: unlock eotsd keyring
 * [#466](https://github.com/babylonlabs-io/finality-provider/pull/466) feat: hot backup eotsd
 
 ## v1.0.0